Refactor DTO Rust and TypeScript renderers
